+++ .gitignore
@@ -1,4 +1,3 @@
 /.vscode/
-*.o
-main
+*.out
