+++ valgs/LinkedList.hpp
@@ -1,13 +1,76 @@
-#include "LinkedList.h"
-#include <optional>
+#ifndef LINKED_LIST_H
+#define LINKED_LIST_H
+
+#include <functional>
 #include <iterator>
+#include <optional>
 
 template <typename T>
-LinkedList<T>::Node::Node(T data)
+class LinkedList
 {
-    Data = data;
-    Next = nullptr;
-}
+    struct Node
+    {
+        T data;
+        Node *next;
+        Node(T data) : data(data), next(nullptr){};
+    };
+
+    struct Iterator
+    {
+        using value_type = T;
+        using element_type = T;
+        using iterator_category = std::forward_iterator_tag;
+
+      private:
+        Node *current;
+
+      public:
+        Iterator();
+        Iterator(Node *node);
+        Iterator(const Iterator &iter);
+        Iterator &operator=(const Iterator &other);
+        T &operator*();
+        const T &operator*() const;
+        Iterator &operator++();
+        Iterator operator++(int);
+        int operator-(const Iterator &other) const;
+        bool operator==(const Iterator &other) const;
+        bool operator!=(const Iterator &other) const;
+    };
+
+  private:
+    Node *_tail;
+    Node *_head;
+    int _count;
+
+  public:
+    void delete_all_nodes();
+    void add_head(T data);   // TODO: test add_head
+    int count();             // TODO: test count
+    std::optional<T> head(); // TODO: test head
+    std::optional<T> tail(); // TODO: test tail
+    void clear();            // TODO: test clear
+    ~LinkedList();
+    LinkedList();
+    Node *get_node_at(int index);       // TODO: get_node_at
+    std::optional<T> get_at(int index); // TODO: test get_at
+    void set_at(int index, T value);    // TODO: test set_at
+    std::optional<T> pop_head();        // TODO: test pop_head
+    std::optional<T> pop_tail();        // TODO: test pop_tail
+    void insert(int index, T data);     // TODO: test insert
+    void add_tail(T data);              // TODO: test add_tail
+    bool remove(T data);                // TODO: remove
+    bool remove_at(int index);          // TODO: test remove_at
+    // TODO: add remove_by version accepting predicate that accepts T by value
+    bool remove_by(std::function<bool(T &)> predicate); // TODO: test remove_by
+    bool contains(T item);                              // TODO: test contains
+
+    // TODO: test iterator
+    Iterator begin();
+    Iterator begin() const;
+    Iterator end();
+    Iterator end() const;
+};
 
 template <typename T>
 LinkedList<T>::Iterator::Iterator()
@@ -30,14 +93,13 @@ LinkedList<T>::Iterator::Iterator(const typename LinkedList<T>::Iterator &iter)
 template <typename T>
 T &LinkedList<T>::Iterator::operator*()
 {
-    return current->Data;
+    return current->data;
 }
 
 template <typename T>
-const T &LinkedList<T>::Iterator::operator*()
-    const
+const T &LinkedList<T>::Iterator::operator*() const
 {
-    return current->Data;
+    return current->data;
 }
 
 template <typename T>
@@ -50,7 +112,7 @@ typename LinkedList<T>::Iterator &LinkedList<T>::Iterator::operator=(const typen
 template <typename T>
 typename LinkedList<T>::Iterator &LinkedList<T>::Iterator::operator++()
 {
-    current = current->Next;
+    current = current->next;
     return *this;
 }
 
@@ -63,22 +125,19 @@ typename LinkedList<T>::Iterator LinkedList<T>::Iterator::operator++(int)
 }
 
 template <typename T>
-int LinkedList<T>::Iterator::operator-(const LinkedList::Iterator &other)
-    const
+int LinkedList<T>::Iterator::operator-(const LinkedList::Iterator &other) const
 {
     return 0;
 }
 
 template <typename T>
-bool LinkedList<T>::Iterator::operator==(const LinkedList::Iterator &other)
-    const
+bool LinkedList<T>::Iterator::operator==(const LinkedList::Iterator &other) const
 {
     return current == other.current;
 }
 
 template <typename T>
-bool LinkedList<T>::Iterator::operator!=(const LinkedList::Iterator &other)
-    const
+bool LinkedList<T>::Iterator::operator!=(const LinkedList::Iterator &other) const
 {
     return current != other.current;
 }
@@ -95,7 +154,7 @@ void LinkedList<T>::add_head(T data)
     }
     else
     {
-        _head->Next = node;
+        _head->next = node;
         _head = node;
     }
     _count++;
@@ -114,7 +173,7 @@ std::optional<T> LinkedList<T>::head()
     {
         return std::nullopt;
     }
-    return _head->Data;
+    return _head->data;
 }
 
 template <typename T>
@@ -124,7 +183,7 @@ std::optional<T> LinkedList<T>::tail()
     {
         return std::nullopt;
     }
-    return _tail->Data;
+    return _tail->data;
 }
 
 template <typename T>
@@ -160,7 +219,7 @@ typename LinkedList<T>::Node *LinkedList<T>::get_node_at(int index)
     Node *current = _tail;
     for (int i = 0; i < index; i++)
     {
-        current = current->Next;
+        current = current->next;
     }
     return current;
 }
@@ -173,7 +232,7 @@ std::optional<T> LinkedList<T>::get_at(int index)
     {
         return std::nullopt;
     }
-    return node->Data;
+    return node->data;
 }
 
 template <typename T>
@@ -184,7 +243,7 @@ void LinkedList<T>::set_at(int index, T value)
     {
         return;
     }
-    node->Data = value;
+    node->data = value;
 }
 
 template <typename T>
@@ -194,7 +253,7 @@ std::optional<T> LinkedList<T>::pop_head()
     {
         return std::nullopt;
     }
-    T last_head_data = _head->Data;
+    T last_head_data = _head->data;
     delete _head;
     if (_count == 1)
     {
@@ -215,7 +274,7 @@ std::optional<T> LinkedList<T>::pop_tail()
     {
         return std::nullopt;
     }
-    T last_tail_data = _tail->Data;
+    T last_tail_data = _tail->data;
     if (_count == 1)
     {
         delete _tail;
@@ -224,7 +283,7 @@ std::optional<T> LinkedList<T>::pop_tail()
         _count--;
         return last_tail_data;
     }
-    Node *new_tail = _tail->Next;
+    Node *new_tail = _tail->next;
     delete _tail;
     _tail = new_tail;
     _count--;
@@ -237,7 +296,7 @@ void LinkedList<T>::insert(int index, T data)
     Node *new_node = new Node(data);
     if (index == 0)
     {
-        add_tail(data);
+        add_tail(new_node);
         return;
     }
     if (index == _count)
@@ -247,8 +306,8 @@ void LinkedList<T>::insert(int index, T data)
     }
 
     Node *previous_node = get_node_at(index - 1);
-    new_node->Next = previous_node->Next;
-    previous_node->Next = new_node;
+    new_node->next = previous_node->next;
+    previous_node->next = new_node;
 
     _count++;
 }
@@ -256,66 +315,96 @@ void LinkedList<T>::insert(int index, T data)
 template <typename T>
 void LinkedList<T>::add_tail(T data)
 {
-    Node *new_node = new Node(data);
+    Node *node = new Node(data);
 
     if (_count == 0)
     {
-        _tail = new_node;
-        _head = new_node;
+        _tail = node;
+        _head = node;
     }
     else
     {
-        new_node->Next = _tail;
-        _tail = new_node;
+        node->next = _tail;
+        _tail = node;
     }
 
     _count++;
 }
 
 template <typename T>
-void LinkedList<T>::remove(T data)
+bool LinkedList<T>::remove(T data)
 {
-    if (_head->Data == data)
+    if (_head->data == data)
     {
         pop_head();
-        return;
+        return true;
     }
-    if (_tail->Data == data)
+    if (_tail->data == data)
     {
         pop_tail();
-        return;
+        return true;
+    }
+    Node *current = _tail;
+
+    for (int i = 0; i <= _count - 2; i++)
+    {
+        Node *previous = current;
+        current = current->next;
+        if (current->data == data)
+        {
+            previous->next = current->next;
+            delete current;
+            _count--;
+            return true;
+        }
+    }
+    return false;
+}
+template <typename T>
+bool LinkedList<T>::remove_by(std::function<bool(T &)> predicate)
+{
+    if (predicate(_head->data))
+    {
+        pop_head();
+        return true;
+    }
+    if (predicate(_tail->data))
+    {
+        pop_tail();
+        return true;
     }
     Node *current = _tail;
     for (int i = 0; i <= _count - 2; i++)
     {
         Node *previous = current;
-        current = current->Next;
-        if (current->Data == data)
+        current = current->next;
+        if (predicate(current->data))
         {
-            previous->Next = current->Next;
+            previous->next = current->next;
             delete current;
             _count--;
-            return;
+            return true;
         }
     }
+    return false;
 }
 
 template <typename T>
-void LinkedList<T>::remove_at(int index)
+bool LinkedList<T>::remove_at(int index)
 {
     if (index >= _count || index < 0)
     {
-        return;
+        return false;
     }
     if (index == 0)
     {
         pop_tail();
-        return;
+        return true;
     }
     if (index == _count - 1)
     {
         pop_head();
-        return;
+        return true;
     }
 
     auto previous_des_node = get_node_at(index - 1);
@@ -323,6 +412,7 @@ void LinkedList<T>::remove_at(int index)
     previous_des_node->Next = desired_node->Next;
     delete desired_node;
     _count--;
+    return true;
 }
 
 template <typename T>
@@ -330,7 +420,7 @@ void LinkedList<T>::delete_all_nodes()
 {
     while (_tail != nullptr)
     {
-        Node *next = _tail->Next;
+        Node *next = _tail->next;
         delete _tail;
         _tail = next;
     }
@@ -343,8 +433,7 @@ typename LinkedList<T>::Iterator LinkedList<T>::begin()
 }
 
 template <typename T>
-typename LinkedList<T>::Iterator LinkedList<T>::begin()
-    const
+typename LinkedList<T>::Iterator LinkedList<T>::begin() const
 {
     return Iterator(_tail);
 }
@@ -356,10 +445,22 @@ typename LinkedList<T>::Iterator LinkedList<T>::end()
 }
 
 template <typename T>
-typename LinkedList<T>::Iterator LinkedList<T>::end()
-    const
+typename LinkedList<T>::Iterator LinkedList<T>::end() const
 {
     return Iterator();
 }
 
-template class LinkedList<int>;
+template <typename T>
+bool LinkedList<T>::contains(T item)
+{
+    auto current = _tail;
+    while (current != nullptr)
+    {
+        if (current->data == item)
+            return true;
+        current = current->next;
+    }
+    return false;
+}
+
+#endif
